Breaking change in naming convention
Use "GrowLights" to access multiple devices and "GrowLight(d3)" for
single ones.

This allows for proper autocomplete, based on types.

Also give error message if accumulation (min/max etc.) or device id is
missing.

+++ src/stationeers_pytrapic/intrinsics.py
@@ -754,5 +754,5 @@ def bnezal(a: Register | float, b: Register | float):
     return "bnezal " + f"{a} {b}"
 
 
-def HASH(name: str) -> float:
+def HASH(name: str) -> int:
     return f"HASH({name})"
